Added song fetching for recorder feature

+++ features/song/impl/src/main/java/com/betocrod/features/song/impl/SongScreen.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.launch
 fun SongScreen(
     viewModel: SongViewModel = hiltViewModel(),
     onBackClick: () -> Unit,
-    onRecordClick: () -> Unit,
+    onRecordClick: (Int) -> Unit,
     onPlayAudio: () -> Unit
 ) {
     val exoplayer = viewModel.player
